+++ app/Http/Controllers/UserController.php
@@ -15,15 +15,19 @@ class UserController extends Controller
 {
     public function index(Request $request)
     {
-        $data = User::orderBy('id')->get();
-        return view('auth.admin.user.index',compact('data'));
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            $data = User::orderBy('id')->get();
+            return view('auth.admin.user.index', compact('data'));
+        }
     }
 
 
     public function create()
     {
-        $roles = Role::pluck('name','name')->all();
-        return view('auth.admin.user.create',compact('roles'));
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            $roles = Role::pluck('name', 'name')->all();
+            return view('auth.admin.user.create', compact('roles'));
+        }
     }
 
 
@@ -32,31 +36,35 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, [
-            'firstname' => 'required',
-            'lastname' => 'required',
-            'email' => 'required|email|unique:users,email',
-            'password' => 'required|same:confirm-password',
-            'roles' => 'required'
-        ]);
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            $this->validate($request, [
+                'firstname' => 'required',
+                'lastname' => 'required',
+                'email' => 'required|email|unique:users,email',
+                'password' => 'required|same:confirm-password',
+                'roles' => 'required'
+            ]);
 
-        $input = $request->all();
-        $input['password'] = Hash::make($input['password']);
+            $input = $request->all();
+            $input['password'] = Hash::make($input['password']);
 
-        $user = User::create($input);
-        $user->assignRole($request->input('roles'));
+            $user = User::create($input);
+            $user->assignRole($request->input('roles'));
 
-        return redirect()->route('users.index')
-            ->with('success','User created successfully');
+            return redirect()->route('users.index')
+                ->with('success', 'User created successfully');
+        }
     }
 
     public function edit($id)
     {
-        $user = User::find($id);
-        $roles = Role::pluck('name','name')->all();
-        $userRole = $user->roles->pluck('name','name')->all();
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            $user = User::find($id);
+            $roles = Role::pluck('name', 'name')->all();
+            $userRole = $user->roles->pluck('name', 'name')->all();
 
-        return view('auth.admin.user.edit',compact('user','roles','userRole'));
+            return view('auth.admin.user.edit', compact('user', 'roles', 'userRole'));
+        }
     }
 
     /**
@@ -64,35 +72,39 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'firstname' => 'required',
-            'lastname' => 'required',
-            'email' => 'required|email|unique:users,email,'.$id,
-            'password' => 'same:confirm-password',
-            'roles' => 'required'
-        ]);
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            $this->validate($request, [
+                'firstname' => 'required',
+                'lastname' => 'required',
+                'email' => 'required|email|unique:users,email,' . $id,
+                'password' => 'same:confirm-password',
+                'roles' => 'required'
+            ]);
 
-        $input = $request->all();
-        if(!empty($input['password'])){
-            $input['password'] = Hash::make($input['password']);
-        }else{
-            $input = Arr::except($input,array('password'));
+            $input = $request->all();
+            if (!empty($input['password'])) {
+                $input['password'] = Hash::make($input['password']);
+            } else {
+                $input = Arr::except($input, array('password'));
+            }
+
+            $user = User::find($id);
+            $user->update($input);
+            DB::table('model_has_roles')->where('model_id', $id)->delete();
+
+            $user->assignRole($request->input('roles'));
+
+            return redirect()->route('users.index')
+                ->with('success', 'User updated successfully');
         }
-
-        $user = User::find($id);
-        $user->update($input);
-        DB::table('model_has_roles')->where('model_id',$id)->delete();
-
-        $user->assignRole($request->input('roles'));
-
-        return redirect()->route('users.index')
-            ->with('success','User updated successfully');
     }
 
     public function destroy($id): \Illuminate\Http\RedirectResponse
     {
-        User::find($id)->delete();
-        return redirect()->route('users.index')
-            ->with('success','User deleted successfully');
+        if (auth()->user()->hasAnyRole('Admin', 'admin')) {
+            User::find($id)->delete();
+            return redirect()->route('users.index')
+                ->with('success', 'User deleted successfully');
+        }
     }
 }
